[Task 1.1] Create MVC Project
# iOS App Initial Setup

## Task [1.1]

- **Target** | iOS 15+
- **Structure** | MVC folders: `Models/`, `Views/`, `Controllers/`, `Services/`, `Resources/`, `Core/`
- **Base** | Setup base classes like `BaseViewController`
- **Build** | Project builds successfully with latest Xcode
- **Config** | Bundle Identifier & Team configured
- **Git** | `.gitignore` optimized for iOS projects

## Tech Stack

**UIKit**, **MVC**, **Application Lifecycle**
