+++ WeatherApp/WeatherApp/Core/SceneDelegate.swift
