+++ WeatherApp/WeatherApp/Core/WeatherApp.xcdatamodeld/WeatherApp.xcdatamodel/contents
