+++ WeatherApp/WeatherApp/Resources/Assets.xcassets/Contents.json
